Moved Labeler Initialization, fixed function names
Realized the Labeler never really needs to be changed for our current scope so moved it outside of the functions as a class variable. made sure to use "label" in MLKitProcess class and "tag" outside of it. added autoLabelImage alongside autoLabelImages depending on the input (was confused by differences in class diagram and api doc for function requirements)