Update MLKit added auto tag to main activity
got rid of unused function overrides, cleaned up MLKit class code, made sure it worked with the updated singlephotoview. Added most of the code that will be needed for auto on device labeling, it is commented out till a bool or something from settings can be used to know when to do it.