adjusted tagging and fixed init gallery view
auto tags do not rotate the image to avoid concurrent write issues, tag suggestions still do this. Moved permissions check before sign in and gallery view data gathering after sign in stuff is done, this fixes the issue of nothing loading in the gallery the first time you open the app. A good amount of MainActivity has comments that can be deleted at this point.