added server tagging check, removed unused code from MLKit
added an if statement to check if server auto tagging is turned on in the main activity right after on device auto tagging is checked for. removed unused functions/code/variables from mlkit class, made functions private where applicable. noticed there is still an error with when we are calling for autotagging to take place, at the moment auto tagging will run before the user has a chance to turn on server auto tagging. which might not really be an issue but seemed like it should be noted.