Overloaded Photo constructor for meta data retrieval
Overloaded Photo Constructor to take only a path, was able to get date/time information from the image.  cant figure how to save it to the database consistently. fixed an issue where photo AutoTag boolean in the database was getting set to false after leaving and re-entering the app, not sure why it wasnt getting reset to true with auto tag still on. might need someone to handle the database calls.