Updated Photo Class to retrieve meta data
Updated Photo Class to retrieve meta data from the image exif information. This included adding the logic for the metadata to be added to the photo object at the correct time. The location information is not yet being retrieved correctly. Fixed an issue where user photo information was being deleted on each app launch.