Thumbnail optimization and photo.rotation
removed filtering and reduced the resolution of the thumbnails being shown in the gallery to reduce the time it takes to load the gallery view. added a "rotation" variable to the photo to reduce the time it takes to display the photo as a thumbnail as well as in single photo view.