Updated Flow and Tracking of Autotagging
Moved autotagging at launch to take place inside the gallery view fragment, used a thread to wait for the autotagging variables to be retrieved for each photo. added separate bool variables to the photo object to indicate if a photo has been server tagged or ML auto tagged. added a db write to the server tagging functions in main to record a server autotagging bool under each photo.